+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 3.8.0 (2025-11-05)
+
+### Added
+
+- **Popover Component**: 
+  - Add `@vueuse/core` dependency;
+  - Add and register Popover, PopoverContent, and PopoverTrigger components in the main index;
+  - Add Popover stories.
+
+# 3.7.1 (2025-11-03)
+
+### Fixed
+
+- **Upload Modal**: Fix close event emitter
+
 # 3.7.0 (2025-10-31)
 
 ### Added